[types] Housekeeping (no functional changes): Move geometry types to own module; Add Size object type; Break out BrushStyle, ShadowStyle to own files; Normalize quote type on import statements.
