+++ src/modules/elements/BarGraph.ts
@@ -1,6 +1,8 @@
 import { evaluateValue } from '../../utils/helpers'
 import { ILayerElement, IValuedElement, RenderContext2D } from '../interfaces'
-import { BrushStyle, ParseState, Rectangle } from '../types'
+import { ParseState } from '../types'
+import { Rectangle } from '../geometry'
+import { BrushStyle } from "./"
 
 // Draws a values series as a basic horizontal bar graph onto a canvas context.
 export default class BarGraph implements ILayerElement, IValuedElement
