+++ src/modules/elements/index.ts
@@ -1,12 +1,15 @@
 // re-exports all contained modules from a single file
 
 export { default as BarGraph } from './BarGraph'
+export { default as BrushStyle } from './BrushStyle'
 export { default as CanvasFilter } from './CanvasFilter'
 export { default as CompositionMode } from './CompositionMode'
 export { default as DrawingStyle } from './DrawingStyle'
 export { default as DynamicImage } from './DynamicImage'
 export { default as LineStyle } from './LineStyle'
 export { default as RoundProgressGauge } from './RoundProgressGauge'
+export { default as ShadowStyle } from './ShadowStyle'
 export { default as StyledRectangle } from './StyledRectangle'
 export { default as StyledText } from './StyledText'
 export { default as Transformation } from './Transformation'
+export { TransformScope } from './Transformation'
