+++ src/modules/geometry/index.ts
@@ -1,0 +1,6 @@
+// re-exports all contained modules from a single file
+
+export { default as Rectangle } from './Rectangle'
+export { default as Vect2d } from './Vect2d'
+export { default as Size } from './Size'  // inherits Vect2d
+export { SizeType } from './Size'
